maint, remove unused findFiles() code (#ifdef USE_FTS)
The plan was to use fts_read() instead of readdir() to avoid calling stat()
for each data file. (We are doing this in --exclude-older-than).

However this unused code is AFAIR not completely correct and also fts_open()
was slow on NFS. If we ever want to do our plan ... just revert this patch.